move peer to p2p file
