+++ package.json
@@ -88,9 +88,11 @@
     "isomorphic-fetch": "^2.2.1",
     "iterall": "^1.2.2",
     "libp2p": "^0.26.2",
+    "libp2p-kad-dht": "^0.16.0",
     "libp2p-mplex": "^0.8.5",
     "libp2p-secio": "^0.11.1",
     "libp2p-tcp": "^0.13.2",
+    "libp2p-webrtc-star": "^0.16.1",
     "lodash": "^4.17.15",
     "node-fetch": "^2.6.0",
     "react": "^16.8.6",
